requirements.txt
torch
torchvision
clip-anytorch
pillow
pandas
streamlit
